// api/src/services/_serviceTypes.ts
export interface BaseCartItem extends Model {
    cart_item_id: number;
    cart_id: number;
    productNo: string;
    thumbnailUrl?: string;
    promotionId?: string;
    quantity: number;
    finalPrice: number;
    reserved: boolean;
}

export interface BaseCart extends Model {
    cart_id: number;
    customer_id: number | null;
}

export interface BaseProductRaw {
    id: number;
    productNo: string;
    productName: string;
    price: number;
    description: string;
    category_id: number;
    subcategory_id: number | null;
    thumbnailUrl: string | null;
    createdAt: Date;
    updatedAt: Date;
    status: string;
}
// ...
export interface JoinReqProduct extends BaseProduct {
    Inventory: BaseInventory;
}

// Flattened product table query with included Inventory table data
export interface RawJoinReqProduct extends BaseProductRaw {
    "Inventory.inventory_id": number;
    "Inventory.product_id": number;
    "Inventory.stock": number;
    "Inventory.reserved": number;
    "Inventory.available": number;
}
// ...
export interface JoinReqCartItem extends BaseCartItem {
    Product: JoinReqProduct;
}

// Cart query with included "CartItem > Product > Inventory"
export interface JoinReqCart extends BaseCart {
    CartItem: JoinReqCartItem[];
}